extract demodata
